Add learned solution to style
Looked at solution for answer on Big Button.
Margin and Padding have 'y' and 'x' axis for
parameters. Example, padding: 16px 8px. This adds
padding of 16px on the top and bottom, and 8px on
sides of element. Example, margin: 0 auto. This
centers the element horizontally while not adding
space vertically.