Add flex solution
Added comments determining which are my code.
Added flex to parent and nested parents so
grand children is affected by flex.
Used mostly gap for spacing between items, and
centered content and alignment on both sides of
header.
Added padding of 8px to heading because I could
not figure out a different way to add spacing
between the content touching the edges of the
header. Gap only adds space between items relative
to parent dimensions, but header does not specific
dimensions (height and width not declared).