Add alignment of texts
I got it to where the icon, header, content, and
buttons should relatively be on the page. I used
flex to take up empty space and make sure the
icon doesn't shrink. I just need the buttons to be
closer together and have the contents not touch
the edges of the modal. It almost looks like the
desired outcome.