Add sidebar
Added spacing between the links and removed margin
and padding of list so it sits on the sidebar
properly. Aligned link to the left side of sidebar
as in the desired outcome png. Sidebar coincides
with self checks and is completed.
Cleaned up style sheet so there's no duplicate
cards and display flex is above what I coded.